\name{colourlovers-package}
\alias{colourlovers-package}
\alias{colourlovers}
\docType{package}
\title{R client for the COLOURlovers API}
\description{Provides access to the COLOURlovers API for accessing colors and color palettes for use in R graphics.}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://www.colourlovers.com/}{COLOURlovers}

\href{http://www.colourlovers.com/api}{COLOURlovers API}
}
\keyword{ package }
%\seealso{}
%\examples{}
