% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clpng.r
\name{clpng}
\alias{clpng}
\alias{clpie}
\alias{plot.clcolor}
\alias{plot.clcolors}
\alias{plot.clpalette}
\alias{plot.clpalettes}
\alias{plot.clpattern}
\alias{plot.clpatterns}
\title{Plot a COLOURlovers object}
\usage{
clpng(x, ...)

clpie(x, ...)

\method{plot}{clcolor}(x, type = "png", ...)
}
\arguments{
\item{x}{A colourlovers color, palette, or pattern object.}

\item{\dots}{Ignored.}

\item{type}{One of \dQuote{png} (the default) or \dQuote{pie}.}
}
\value{
The \code{x} object is invisibly returned.
}
\description{
Plot a color, palette, or pattern color pie or PNG image in a plotting
device
}
\details{
Plot a colourlovers PNG image in a new plotting window.
}
\examples{
e <- function(e) NULL # function for tryCatch to fail in examples

# plot a color clpng
co <- tryCatch( clcolor(rgb(0,0,1)), error = e)
if(!is.null(co)) plot(co)

# plot a pattern clpng
p <- tryCatch( clpattern('1451'), error = e)
if(!is.null(p)) plot(p)
if(!is.null(p)) clpng(p)

# plot colors from a palette
p <- tryCatch( clpalette('113451'), error = e)
if(!is.null(p)) plot(p, type='pie')
if(!is.null(p)) clpie(p)


\dontrun{
# plot multiple palettes (interactively)
p <- tryCatch( clpalettes('top'), error = e)
if(!is.null(p)) plot(p) # PNG images
if(!is.null(p)) plot(p, type='pie') # pie chart swatches
}

}
\author{
Thomas J. Leeper
}
