% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourlovers-package.R
\docType{package}
\name{colourlovers-package}
\alias{colourlovers-package}
\alias{colourlovers}
\title{R client for the COLOURlovers API}
\description{
Provides access to the COLOURlovers API for accessing colors and color
palettes for use in R graphics.
}
\author{
Thomas J. Leeper
}
\keyword{package}
