% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatch.r
\name{swatch}
\alias{swatch}
\title{Extract colors from an object}
\usage{
swatch(x, ...)
}
\arguments{
\item{x}{A colourlovers color, palette, or pattern object.}

\item{...}{Ignored.}
}
\value{
A list of character vectors containing hexidecimal representations
  of colors.
}
\description{
Extract a list of colors from a COLOURlovers object
}
\details{
Extract color(s) from a colourlovers object.
}
\examples{
\dontrun{
# Get top colors
swatch(clcolors('top'))

# Get colors from a specific palette
swatch(clpalette('113451'))

# Get colors from specific pattern
swatch(clpattern('1451'))
}
}
\author{
Thomas J. Leeper
}
